move Groups Pager into Components as Common Components
